requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-generativeai==0.3.2
firebase-admin==6.3.0
Werkzeug==3.0.1
python-dotenv==1.0.0
WeasyPrint==59.0
pydyf==0.6.0   
pexels-api==1.0.1
razorpay==1.3.0
setuptools==69.5.1
